organizer: split csharp analyzer tests
